remove relative path tricks
