+++ sqrl-tools/sqrl-cli/src/test/resources/dependency/main1.sqrl
