+++ sqrl-tools/sqrl-cli/src/test/resources/dependency/main2.sqrl
