+++ sqrl-tools/sqrl-cli/src/test/resources/dependency/packageWDependencies.json
