+++ sqrl-tools/sqrl-cli/src/test/resources/dependency/packageWODependencies.json
