+++ sqrl-tools/sqrl-cli/src/test/resources/snapshots/com/datasqrl/cmd/PackageBootstrapTest/testRetailPackaging.txt
@@ -209,7 +209,6 @@ package.json
 build.gradle
 c360-full-graphqlv1
 c360-full-graphqlv1/schema.graphqls
-c360-full.sqrl
 ecommerce-avro
 ecommerce-avro/orders.avsc
 ecommerce-avro/orders.table.json
@@ -238,7 +237,7 @@ package.json
   },
   "script" : {
     "graphql" : "c360-full-graphqlv1/schema.graphqls",
-    "main" : "c360-full.sqrl"
+    "main" : "c360-full-graphqlv1/schema.graphqls"
   },
   "package" : {
     "name" : "datasqrl.examples.Retail",
